% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_cov.R
\name{cov}
\alias{cov}
\title{Coverage / C-metric of two (approximation) sets}
\usage{
cov(x, y, ...)
}
\arguments{
\item{x}{[\code{matrix}]\cr
First numeric matrix of points (each colum contains one point).}

\item{y}{[\code{matrix}]\cr
Second numeric matrix of points (each colum contains one point).}

\item{...}{[any]\cr
Not used.}
}
\value{
Coverage value.
}
\description{
Calculates the Coverage metric [1] (also known as the C-metric) given two sets
of points.
}
\details{
Given two (approximation) sets \eqn{X = \{x_1, \ldots, x_{|X|}\}}
and \eqn{Y = \{y_1, \ldots, y_{|Y|}\}} this indidcator calculates the fraction
of points from \eqn{Y} which are dominated by at least one point \eqn{x \in X}.
Formally, it is defined as:
\deqn{
  C(X, Y) = \frac{|\{y \in Y \mid \exists x \in X,\, x \preceq y\}|}{|Y|}.
}
It holds that \eqn{C(X, Y) \in [0,1]}. Note that in general \eqn{C(X, Y) \neq 1 - C(Y, X)}.
}
\note{
Keep in mind that this function assumes all objectives to be minimized.
In case at least one objective is to be maximized, the data needs
to be transformed accordingly in advance.
}
\references{
[1] E. Zitzler, Evolutionary algorithms for multiobjective optimization:
Methods and applications, Ph.D. thesis,  Swiss Federal Institute of
Technology Zurich (1999).
}
\seealso{
Other multi-objective performance indicators: 
\code{\link{df_get_indicators}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{hv}()},
\code{\link{os}()},
\code{\link{r1}()},
\code{\link{rse}()}
}
\concept{multi-objective performance indicators}
\keyword{optimize}
