% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_df_get_indicators.R
\name{df_get_indicators}
\alias{df_get_indicators}
\title{Calculate multi-objective performance indicators}
\usage{
df_get_indicators(x, obj.cols, unary, format = "long")
}
\arguments{
\item{x}{[\code{data.frame}]\cr
Input data frame. Requires at least the column names given by \code{obj.cols}.}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective functions.
Default is \code{c("f1", "f2")}.}

\item{unary}{[\code{list}]\cr
Named list of indicators. The names must be strings which correspond to
the function name of the indicator (e.g. \dQuote{gd} for \code{\link{gd}}).
The value is an (empty) named list of parameter values for the indicator
(e.g. \code{list(p = 2)} to modify the \eqn{p} parameter for GD).}

\item{format}{[\code{character(1)}]\cr
If \dQuote{long}, the data is returned as a data frame in long format. I.e.,
there is a column \dQuote{indicator} and another column \dQuote{value} for
the corresponding indicator values. This format is the default and particulary
helpful for visualization with \pkg{ggplot2}.
In contrast, for format \dQuote{wide}, there is one column for each
indicator. This format is less redundant and memory-intensive.}
}
\value{
[\code{data.frame}] A data frame with columns \dQuote{problem},
  \dQuote{algorithm}, \dQuote{repl} and columns with the respective
  indicator values (see argument \code{format} for details).
}
\description{
The function expects a data frame with Pareto approximation sets given
in the columns passed by \code{obj.cols} and optional columns \dQuote{problem}
\dQuote{algorithm} and \dQuote{repl}. Given a set of multi-objective performance
indicators, the function splits the data by the meta-column and calculates the
indicator values for each approximation set.
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{r1}()},
\code{\link{rse}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other multi-objective performance indicators: 
\code{\link{cov}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{hv}()},
\code{\link{os}()},
\code{\link{r1}()},
\code{\link{rse}()}
}
\concept{mootools}
\concept{multi-objective performance indicators}
\keyword{optimize}
