% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_domination.R
\name{dominates}
\alias{dominates}
\alias{does_dominate}
\alias{is_dominated}
\alias{\%dominates\%}
\alias{\%is_dominated\%}
\title{Dominance relation checks}
\usage{
dominates(x, y)

does_dominate(x, y)

is_dominated(x, y)

x \%dominates\% y

x \%is_dominated\% y
}
\arguments{
\item{x}{[\code{numeric}]\cr
First vector.}

\item{y}{[\code{numeric}]\cr
Second vector.}
}
\value{
[\code{logical(1)}]
}
\description{
Check if a vector dominates another (\code{dominates}) or is
dominated by another (\code{is_dominated}). There are corresponding infix
operators \code{dominates} and \code{is_dominated}.
}
\seealso{
Other pareto-dominance: 
\code{\link{dominated}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other mootools: 
\code{\link{dominated}()},
\code{\link{get_nadir}()},
\code{\link{nds}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}
}
\concept{mootools}
\concept{pareto-dominance}
\keyword{optimize}
