% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookkeeping_logger.R
\name{log_init}
\alias{log_init}
\title{Logger initialization}
\usage{
log_init(what, init.size, at = NULL, to = NULL)
}
\arguments{
\item{what}{[\code{named(character)}]\cr
Named character vector. The names (i.e. \code{names(what)}) describe which
elements we want to log and the actual values represent the types/modes these
values are stored as. Possible values are \dQuote{integer}, \dQuote{numeric},
\dQuote{character}, \dQuote{logical} for the respective \strong{scalar}
atomic values or \dQuote{list} for any other complex R objects, e.g., matrices,
lists etc.
Note that internally \code{c("iter" = "numeric")} is attached automatically
and \code{iter} must be passed to \code{\link{log_update}} in subsequent
calls.}

\item{init.size}{[\code{integer(1)}]\cr
Initial size of the log.}

\item{at}{[\code{integer}]\cr
Optional integer vector of times when logging should take place.
If the iteration counter is in \code{at}, the state is logged, otherwise it
is not.}

\item{to}{[\code{character(1)}]\cr
File path to log-directory.
If not \code{NULL}, each call to \code{log_update} stores the serialized
state in file \dQuote{<to>/<iteration>.rds}.}
}
\value{
[\code{enviroment}] Logger environment.
}
\description{
Generates a fairly flexible container to log stuff.
}
\examples{
log = log_init(c("P" = "list", "f" = "list", "C" = "character"),
  init.size = 5L, at = c(1, 3))

# note that div is not subject to logging since it was not specified in log_init
log_update(log, P = matrix(runif(10), ncol = 2L), C = "a",
  f = runif(10), iter = 1, div = letters[1:3])
log_update(log, P = matrix(runif(10), ncol = 2L), C = "b",
  f = runif(10), iter = 2, div = letters[1:3])
log_update(log, P = matrix(runif(10), ncol = 2L), C = "a",
  f = runif(10), iter = 3, div = letters[1:3])
print(log$df)
}
\seealso{
log_update
}
