% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_nds.R
\name{nds}
\alias{nds}
\title{Fast non-dominated sorting}
\usage{
nds(x)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Numeric matrix of points. Each column contains one point.}
}
\value{
[\code{list}]
  List with the following components
  \describe{
    \item{ranks}{Integer vector of ranks of length \code{ncol(x)}. The higher
    the rank, the higher the domination front the corresponding point is
    located on.}
    \item{dom.counter}{Integer vector of length \code{ncol(x)}. The \eqn{i}th element
    is the domination number of the \eqn{i}th point.}
  }
}
\description{
Fast non-dominated sorting algorithm proposed by Deb et al. [1].
Non-dominated sorting expects a set of points and returns a
set of non-dominated fronts. In short words this is done as follows: the
non-dominated points of the entire set are determined and assigned rank 1.
Afterwards all points with the current rank are removed, the rank is increased
by one and the procedure starts again. This is done until the set is empty, i.e.,
each point is assigned a rank.
}
\note{
This procedure is the key survival selection of the famous NSGA-II multi-objective
evolutionary algorithm [1].
}
\references{
[1] Deb, K., Pratap, A., and Agarwal, S. A Fast and Elitist Multiobjective Genetic
Algorithm: NSGA-II. IEEE Transactions on Evolutionary Computation, 6 (8) (2002),
182-197.
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{df_get_indicators}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{r1}()},
\code{\link{rse}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}
}
\concept{mootools}
\keyword{optimize}
