% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_eaf.R
\name{plot_eaf}
\alias{plot_eaf}
\title{Empirical attainment function (EAF)}
\usage{
plot_eaf(df, obj.cols, percentiles = c(0, 50, 75, 100))
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{problem} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective function values.
Default is \code{c("y1", "y2")}.}

\item{percentiles}{[\code{numeric}]\cr
Percentiles of the EAF that will be plotted as attainment surfaces.}
}
\value{
[\code{\link[ggplot2]{ggplot}}] A ggplot object.
}
\description{
The following description closely follows the excellent introduction into
performance assessment of multi-objective optimizers by Knowles et al. [3].
(Evolutionary) multi-objective algorithms are stochastic. Hence, naturally,
the result produced by such an algorithm can be described by a probability
distribution. This distribution can be descibed by a random set
\deqn{
  Z = \{z^j \in R^m \,|\, j = 1, \ldots, |Z|\}
}
where the size \eqn{|Z|} is also random. The \emph{attainment function}
\eqn{\alpha_Z : R^m \to [0,1]} is defined as
\deqn{
  \alpha_Z(z)
  = P(Z \preceq \{z\}) = P\left(z^1 \preceq z \lor \ldots \lor z^{|Z|} \preceq z\right).
}
where \eqn{z^j \preceq z} denotes that \eqn{z^j} weakly dominates \eqn{z}.
This can be interpreted as the \emph{probability to reach goal} \eqn{z} in
the sense that there is at least one objective vector in the solution set
that weakly dominates, i.e., \dQuote{attains}, \eqn{z}.

Given \eqn{r} independent runs of a stochastic multi-objective optimizer and
denoting by \eqn{X^i, 1 \leq i \leq r} the corresponding Pareto-front
approximation set the \emph{empirical attainment function (EAF)} is defined
as
\deqn{
  \hat{\alpha}_r(z) = \frac{1}{r} \sum_{i=1}^{r} I(X^i \preceq \{z\})
}
where \eqn{I(\cdot)} is the \emph{indictor function} evaluating to \eqn{1}
if and only if the condition is true; \eqn{0} otherwise.
This is a straight-forward estimation from empirical data and simply describes
the frequency of attaining \eqn{z} within \eqn{r} runs.

The EAF can be used for visualization. Here, we plot the so-called
\eqn{k\%}-attainment surface which splits \dQuote{the goals that have been attained
and the goals that have not been attained with a frequency of at least
\eqn{k} percent} [3].

For further details we refer the reader to the technical report by Knowles,
Thiele and Zitzler (reference [3]).
}
\references{
[1] V. Grunert da Fonseca and C. M. Fonseca, The attainment-function approach
to stochastic multiobjective optimizer assessment and comparison, in
Experimental Methods for the Analysis of Optimization Algorithms
(T. Bartz-Beielstein, M. Chiarandini, L. Paquete, and M. Preuss, eds.),
ch. 5, pp. 103-130, Springer Berlin Heidelberg, 2010.

[2] Manuel López-Ibáñez, Luís Paquete, and Thomas Stützle. Exploratory Analysis
of Stochastic Local Search Algorithms in Biobjective Optimization. In T.
Bartz-Beielstein, M. Chiarandini, L. Paquete, and M. Preuss, editors,
Experimental Methods for the Analysis of Optimization Algorithms,
pages 209–222. Springer, Berlin, Germany, 2010. doi: 10.1007/978-3-642-02538-9_9

[3] Knowles, J. D., Thiele, L. and Zitzler, E. A tutorial on the performance
assessment of stochastive multiobjective optimizers. TIK-Report No. 214, Computer
Engineering and Networks Laboratory, ETH Zurich, February 2006 (Revised version.
First version, January 2005). doi: 10.3929/ethz-b-000023822
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{df_get_indicators}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{r1}()},
\code{\link{rse}()},
\code{\link{set_dominates}()},
\code{\link{solow_polasky}()},
\code{\link{which_dominated}()}

Other visualization: 
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()}
}
\concept{mootools}
\concept{visualization}
