% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_eaf_diff.R
\name{plot_eaf_diff}
\alias{plot_eaf_diff}
\title{Empirical attainment function (EAF) difference}
\usage{
plot_eaf_diff(df1, df2, obj.cols, intervals = 5L, percentiles = c(0, 50, 75))
}
\arguments{
\item{df1}{[\code{data.frame}]\cr
First approximation set: a data frame with columns at least those given via
parameter \code{obj.cols}, and \dQuote{repl}. All other columns will be ignored!}

\item{df2}{[\code{data.frame}]\cr
First approximation set: a data frame with columns at least those given via
\code{obj.cols}, and \dQuote{repl}. All other columns will be ignored!}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective function values.
Default is \code{c("y1", "y2")}.}

\item{intervals}{[\code{integer(1)}]\cr
The absolute range of the differences \eqn{[0, 1]} is partitioned into the
number of intervals provided.
Default is 5.}

\item{percentiles}{[\code{numeric}]\cr
Percentiles of the EAF that will be plotted as attainment surfaces.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot the differences between the empirical attainment functions (EAFs) of two
approximations sets side by side adopting \CRANpkg{ggplot2}. The left plot
shows the the values of the left EAF minus the right EAF and the right side
shows it the other way around.
}
\note{
If the given approximation sets are huge in size, generating the plot may take some
time. Here, it is most often helpful to store the plot first in a file. Due
to the huge number of points, the size of the file may become also very large. Here,
it is advisible to save the file in an image file format (PNG or JPEG) instead
of PDF.

This function re-implements partial functionality of \code{\link[eaf]{eafdiffplot}}
realizing the plotting with \CRANpkg{ggplot2} rather than base R plot functions.
}
\examples{
\dontrun{
data(emoas_on_zdt)
df1 = emoas_on_zdt[emoas_on_zdt$algorithm == "nsga2"
  & emoas_on_zdt$problem == "zdt1_2d_2o", ]
df2 = emoas_on_zdt[emoas_on_zdt$algorithm == "smsemoa"
  & emoas_on_zdt$problem == "zdt1_2d_2o", ]
plot_eaf_diff(df1, df2, obj.cols = c("y1", "y2"))
}
}
\references{
[1] V. Grunert da Fonseca and C. M. Fonseca, The attainment-function approach
to stochastic multiobjective optimizer assessment and comparison, in
Experimental Methods for the Analysis of Optimization Algorithms
(T. Bartz-Beielstein, M. Chiarandini, L. Paquete, and M. Preuss, eds.),
ch. 5, pp. 103-130, Springer Berlin Heidelberg, 2010.

[2] Manuel López-Ibáñez, Luís Paquete, and Thomas Stützle. Exploratory Analysis
of Stochastic Local Search Algorithms in Biobjective Optimization. In T.
Bartz-Beielstein, M. Chiarandini, L. Paquete, and M. Preuss, editors,
Experimental Methods for the Analysis of Optimization Algorithms,
pages 209–222. Springer, Berlin, Germany, 2010. doi: 10.1007/978-3-642-02538-9_9.

[3] Knowles, J. D., Thiele, L. and Zitzler, E. A tutorial on the performance
assessment of stochastive multiobjective optimizers. TIK-Report No. 214, Computer
Engineering and Networks Laboratory, ETH Zurich, February 2006 (Revised version.
First version, January 2005). doi: 10.3929/ethz-b-000023822.
}
\seealso{
plot_eaf

Other multi-objective visualizations: 
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()}
}
\concept{multi-objective visualizations}
