% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap visualization}
\usage{
plot_heatmap(df, obj.cols = c("y1", "y2"))
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{prob} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective functions.
Default is \code{c("f1", "f2")}.}
}
\value{
[\code{\link[ggplot2]{ggplot}}] ggplot object.
}
\description{
Visualize a Pareto-front approximation by means of a heatmap.
Here, on the x-axis the objectives are given in the order provided by
\code{obj.cols}. Tthe y-axis shows the different solutions. The color of the
cells indicates the objective value of the respective (objective, solution)
pair. Works with any kind of objectives.
}
