% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.pcp.R
\name{plot_pcp}
\alias{plot_pcp}
\title{Parallel coordinate plot}
\usage{
plot_pcp(df, obj.cols = c("y1", "y2"))
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{prob} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective functions.
Default is \code{c("f1", "f2")}.}
}
\value{
[\code{\link[ggplot2]{ggplot}}] ggplot object.
}
\description{
Visualizes a Pareto-front approximation with parallel coordinate
plots. The x-axis shows the objectives as an ordered factor where the order
is determined by the \code{obj.cols} argument. On the y-axis objective values are
shown colored by solution as lines. Suitable for any number of objectives.
}
