% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Plot radar plot}
\usage{
plot_radar(df, obj.cols = c("y1", "y2"))
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{prob} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective functions.
Default is \code{c("f1", "f2")}.}
}
\value{
[\code{\link[ggplot2]{ggplot}}] ggplot object.
}
\description{
A radar plot or spider plot is an elegant method to visualize
solution quality of solutions with more than two objectives. Essentially it
shows multiple numeric values starting from the same origin in a circular
two-dimensional representation

Users who are into gaming might know the kind of visualization from games like Pro
Evolution Soccer, NHL or most likely any kind of sport games. Here, the
individual strength and weaknesses of players become obvious.

The drawback of this method that it is difficult to compare many solutions
or - even worse - many solutions from different algorithms to get the big picture.
}
