% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_scatter3d.R
\name{plot_scatter3d}
\alias{plot_scatter3d}
\title{3D scatter plot}
\usage{
plot_scatter3d(
  df,
  obj.cols = c("y1", "y2", "y3"),
  max.in.row = 4L,
  package = "scatterplot3d",
  ...
)
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{problem} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 3)}]\cr
Column names of the objective functions.
Default is \code{c("y1", "y2", "y3")}.}

\item{max.in.row}{[\code{integer(1)}]\cr
Maximum number of plots to be displayed side by side in a row.
Default is 4.}

\item{package}{[\code{character(1L)}]\cr
Which package to use for 3d scatterplot generation?
Possible choices are \dQuote{scatterplot3d} (\pkg{scatterplot3d}), \dQuote{plot3D},
(\pkg{plot3D}), \dQuote{plot3Drgl} (\pkg{plot3Drgl}) or \dQuote{plotly} (\pkg{plotly}).
Default is \dQuote{scatterplot3d}.}

\item{...}{[any]\cr
Further arguments passed down to scatterplot function.}
}
\value{
Nothing (function has side-effects by calling the respective drawing routines).
}
\description{
Given a data frame with the results of (multiple) runs of (multiple)
different three-objective optimization algorithms on (multiple) problem instances
the function generates 3D scatterplots of the obtained Pareto-front approximations.
}
\references{
[1] T. Tušar and B. Filipič, "Visualization of Pareto Front Approximations in
Evolutionary Multiobjective Optimization: A Critical Review and the Prosection
Method," in IEEE Transactions on Evolutionary Computation, vol. 19, no. 2,
pp. 225-245, April 2015, doi: 10.1109/TEVC.2014.2313407
}
\seealso{
Other multi-objective visualizations: 
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()}
}
\concept{multi-objective visualizations}
