% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_reference_points.R
\name{get_nadir}
\alias{get_nadir}
\alias{get_ideal}
\title{Reference point approximation}
\usage{
get_nadir(..., sets = NULL)

get_ideal(..., sets = NULL)
}
\arguments{
\item{...}{[\code{matrix}]\cr
Arbirary number of matrizes.}

\item{sets}{[\code{list}]\cr
List of matrizes. This is an alternative way of passing the sets. Can be used
exclusively or combined with \code{...}.}
}
\value{
Reference point (numeric vector).
}
\description{
Helper functions to compute nadir or ideal point from multiple sets of
points, e.g., multiple approximation sets produced by some multi-objective
evolutionary algorithm.
}
\seealso{
Other mootools: 
\code{\link{nds}()}
}
\concept{mootools}
