% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_r.R
\name{r1}
\alias{r1}
\alias{r2}
\alias{r3}
\title{R{1,2,3}-indicators}
\usage{
r1(x, y, ip = NULL, np = NULL, lambda = NULL, utility = "tschebycheff", ...)

r2(x, y, ip = NULL, np = NULL, lambda = NULL, utility = "tschebycheff", ...)

r3(x, y, ip = NULL, np = NULL, lambda = NULL, utility = "tschebycheff", ...)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Point set in column major format.}

\item{y}{[\code{matrix}]\cr
Reference point set in column major format.}

\item{ip}{[\code{numeric}]\cr
The utopia point of the true Pareto-front, i.e., each component of the point
contains the best value if the other objectives are neglected.
If \code{NULL} (default) the ideal point of the union of \code{x} and
\code{y} is used.}

\item{np}{[\code{numeric}]\cr
Nadir point of the true Pareto front or an approximation.
If \code{NULL} (default) the nadir point of the union of \code{x} and
\code{y} is used.}

\item{lambda}{[\code{integer(1)}]\cr
Number of weight vectors to use in estimating the utility function.}

\item{utility}{[\code{character(1)}]\cr
Name of the utility function to use. Must be one of \dQuote{weighted-sum},
\dQuote{tschebycheff} or \dQuote{augmented-tschbycheff}.}

\item{...}{[any]\cr
Not used.}
}
\description{
...
}
\references{
[1] M. P. Hansen and A. Jaszkiewicz. 1998. Evaluating the quality of
approximations to the nondominated set. Imm-rep-1998-7. Institute of
Mathematical Modeling, Technical University of Denmark.
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{df_get_indicators}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{rse}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other multi-objective performance indicators: 
\code{\link{cov}()},
\code{\link{df_get_indicators}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{hv}()},
\code{\link{os}()},
\code{\link{rse}()}
}
\concept{mootools}
\concept{multi-objective performance indicators}
\keyword{optimize}
