% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_domination.R
\name{set_dominates}
\alias{set_dominates}
\title{Check if one set is better than another.}
\usage{
set_dominates(x, y)
}
\arguments{
\item{x}{[\code{matrix}]\cr
First set of points.}

\item{y}{[\code{matrix}]\cr
Second set of points.}
}
\value{
[\code{logical(1)}]
}
\description{
The function checks, whether every point from the second set of points
is dominated by at least one point from the first set.
}
\seealso{
Other pareto-dominance: 
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{which_dominated}()}

Other mootools: 
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{get_nadir}()},
\code{\link{nds}()},
\code{\link{which_dominated}()}
}
\concept{mootools}
\concept{pareto-dominance}
