% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{update_logger}
\alias{update_logger}
\title{Update logger.}
\usage{
update_logger(log, log.stuff = list(), ...)
}
\arguments{
\item{log}{[\code{enviroment}]\cr
Logger initialized with \code{\link{init_logger}}.}

\item{log.stuff}{[\code{list}]\cr
Named list of objects.
The subset specified by argument names of parameter \code{what} of \code{\link{init_logger}}
will be stored.}

\item{...}{[any]\cr
Variable argument list as another possibility to pass objects.
These arguments are merged with \code{log.stuff}. Note that the dot-args
have precdence, i.e., objects in \code{log.stuff} are owerwritten by objects
named equally in the dot-args list.}
}
\value{
Nothing. The function modifies \code{log} in-place.
}
\description{
Store stuff in logger object.
}
\seealso{
init_logger
}
