% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_domination.R
\name{which_dominated}
\alias{which_dominated}
\alias{which_nondominated}
\alias{is_maximally_dominated}
\title{D
etermine which points of a set are (non)dominated}
\usage{
which_dominated(x)

which_nondominated(x)

is_maximally_dominated(x)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Numeric \eqn{(m \times n)} matrix where \eqn{n} is the number of points and \eqn{m}
is the number of objectives.}
}
\value{
Integer vector of positions of (non)dominated points.
}
\description{
Given a matrix with one point per column \code{which_dominated} returns the
column numbers of the dominated points and \code{which_nondominated} the column
numbers of the nondominated points. Function \code{is_maximally_dominated} returns
a logical vector with \code{TRUE} for each point which is located on the last
non-domination level.
}
\examples{
data(mtcars)
# assume we want to maximize horsepower and minimize gas consumption
cars = mtcars[, c("mpg", "hp")]
cars$hp = -cars$hp
idxs = which_nondominated(t(as.matrix(cars)))
\dontrun{
print(mtcars[idxs, ])
plot(cars)
points(cars[idxs, ], col = "tomato", pch = 10)
}
}
\seealso{
Other Pareto-dominance checks: 
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{set_dominates}()}

Other multi-objective tools: 
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{set_dominates}()}
}
\concept{Pareto-dominance checks}
\concept{multi-objective tools}
\keyword{optimize}
