% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_cov.R
\name{cov}
\alias{cov}
\title{Coverage / C-metrix of two (approximation) sets.}
\usage{
cov(x, y)
}
\arguments{
\item{x}{[\code{matrix}]\cr
First point set in column major format.}

\item{y}{[\code{matrix}]\cr
Second point set in column major format.}
}
\value{
[\code{numeric(1)}] Coverage value.
}
\description{
Given two (approximation) sets \eqn{X = \{x_1, \ldots, x_{|X|}\}}
and \eqn{Y = \{y_1, \ldots, y_{|Y|}\}} this indidcator calculates the fraction
of points from \eqn{Y} which are dominated by at least one point \eqn{x \in X}.
Formally it is defined as:
\deqn{
  C(X, Y) = \frac{|\{y \in Y \mid \exists x \in X with x \preceq y\}|}{|Y|}.
}
It holds that \eqn{C(X, Y) \in [0,1]}. Note that in general
\deqn{
  C(X, Y) \neq 1 - C(Y, X).
}
}
\references{
[1] E. Zitzler, Evolutionary algorithms for multiobjective optimization:
Methods and applications, Ph.D. thesis,  Swiss Federal Institute of
Technology Zurich (1999).
}
\seealso{
Other mootools: 
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{rse}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other multi-objective performance indicators: 
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{hv}()},
\code{\link{os}()},
\code{\link{rse}()}
}
\concept{mootools}
\concept{multi-objective performance indicators}
\keyword{optimize}
