% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_domination.R
\name{dominated}
\alias{dominated}
\alias{nondominated}
\title{Check for Pareto-dominance.}
\usage{
dominated(x)

nondominated(x)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Numeric (d x n) matrix where d is the number of objectives and n is the
number of points.}
}
\value{
[\code{logical}]
}
\description{
These functions take a numeric matrix \code{x} where each column corresponds to
a point and return a logical vector. The i-th position of the latter is
\code{TRUE} if the i-th point is dominated by at least one other point for
\code{dominated} and \code{FALSE} for \code{nonDominated}.
}
\seealso{
Other pareto-dominance: 
\code{\link{dominates}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other mootools: 
\code{\link{cov}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}
}
\concept{mootools}
\concept{pareto-dominance}
\keyword{optimize}
