% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_domination.R
\name{dominates}
\alias{dominates}
\alias{does_dominate}
\alias{is_dominated}
\alias{\%dominates\%}
\alias{\%is_dominated\%}
\title{Dominance relation checks}
\usage{
dominates(x, y)

does_dominate(x, y)

is_dominated(x, y)

x \%dominates\% y

x \%is_dominated\% y
}
\arguments{
\item{x}{[\code{numeric}]\cr
First vector.}

\item{y}{[\code{numeric}]\cr
Second vector.}
}
\value{
Single logical value.
}
\description{
Check if a vector dominates another (\code{dominates} or \code{does_dominate})
or is dominated by another (\code{is_dominated}). There are corresponding infix
operators \code{dominates} and \code{is_dominated}.
}
\details{
Given two vectors \eqn{x,y \in R^m} we say that \eqn{x} dominates \eqn{y},
denoted as \eqn{x \preceq y} if and only if
\deqn{
  x_i \leq y_i \, \forall i \in \{1, \ldots, m\}
}
and
\deqn{
  \exists j \in \{1, \ldots, m\}: x_i < y_i.
}
Informally, \eqn{x} dominates \eqn{y} if \eqn{x} is no worse than \eqn{y} in
all components and \eqn{x} is strictly better than \eqn{y} in at least one
component. Note that this definition focuses on minimization of all
objectives. This is no restriction since the maximization of some function
is equivalent to the minimization of its negative counterpart.
}
\examples{
dominates(c(2, 3), c(4, 5))
dominates(c(2, 3), c(2, 3))
is_dominated(c(2, 3), c(4, 5))
is_dominated(c(4, 5), c(2, 3))
c(1, 2, 3) \%dominates\% c(4, 5, 6)
c(1, 2, 3) \%is_dominated\% c(4, 5, 6)
}
\seealso{
Other Pareto-dominance checks: 
\code{\link{dominated}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other multi-objective tools: 
\code{\link{dominated}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}
}
\concept{Pareto-dominance checks}
\concept{multi-objective tools}
\keyword{optimize}
