% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_domination.R
\name{dominates}
\alias{dominates}
\alias{does_dominate}
\alias{is_dominated}
\alias{\%dominates\%}
\alias{\%is_dominated\%}
\title{Dominance relation checks}
\usage{
dominates(x, y)

does_dominate(x, y)

is_dominated(x, y)

x \%dominates\% y

x \%is_dominated\% y
}
\arguments{
\item{x}{[\code{numeric}]\cr
First vector.}

\item{y}{[\code{numeric}]\cr
Second vector.}
}
\value{
[\code{logical(1)}]
}
\description{
Check if a vector dominates another (\code{dominates} or \code{does_dominate})
or is dominated by another (\code{is_dominated}). There are corresponding infix
operators \code{dominates} and \code{is_dominated}.
}
\examples{
dominates(c(2, 3), c(4, 5))
dominates(c(2, 3), c(2, 3))
is_dominated(c(2, 3), c(4, 5))
is_dominated(c(4, 5), c(2, 3))
c(1, 2, 3) \%dominates\% c(4, 5, 6)
c(1, 2, 3) \%is_dominated\% c(4, 5, 6)
}
\seealso{
Other Pareto-dominance checks: 
\code{\link{dominated}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other multi-objective tools: 
\code{\link{dominated}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}
}
\concept{Pareto-dominance checks}
\concept{multi-objective tools}
\keyword{optimize}
