% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_eps.R
\name{eps}
\alias{eps}
\title{\eqn{\varepsilon}-indicator}
\usage{
eps(x, y)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Point set in column major format.}

\item{y}{[\code{matrix}]\cr
Reference point set in column major format.}
}
\value{
[\code{numeric(1)}] Scalar indicator value.
}
\description{
...
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other multi-objective performance indicators: 
\code{\link{cov}()},
\code{\link{gd}()},
\code{\link{hv}()},
\code{\link{os}()}
}
\concept{mootools}
\concept{multi-objective performance indicators}
\keyword{optimize}
