% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_gd.R
\name{gd}
\alias{gd}
\alias{igd}
\alias{gdp}
\alias{igdp}
\alias{ahd}
\title{(Inverted) Generational Distance and Average Hausdorff Distance}
\usage{
gd(x, y, p = 2, modified = TRUE, ...)

igd(x, y, p = 2, modified = TRUE, ...)

gdp(x, y, p = 2, modified = TRUE, ...)

igdp(x, y, p = 2, modified = TRUE, ...)

ahd(x, y, p = 2, modified = TRUE, ...)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Numeric matrix of points (each colum contains one point).}

\item{y}{[\code{matrix}]\cr
The reference set as a numeric matrix of points (each colum contains one point).}

\item{p}{[\code{numeric(1)}]\cr
Parameter \eqn{p} (see description).}

\item{modified}{[\code{logical(1)}]\cr
Should the modified GD/IGD calculation by Schuetze et al. [2] be used?
Default is \code{TRUE}.}

\item{...}{[any]\cr
Not used.}
}
\value{
Single numeric indicator value.
}
\description{
These functions expect as mandatory arguments a point set \eqn{X = \{x_1, \ldots, x_{|X|}\}}
(parameter \code{x}) and  and a set of reference points
\eqn{Y = \{y_1, \ldots, y_{|Y|}\}} (parameter \code{y}). For details on the
optional argument \code{p} and \code{modified} see the section on details.
\itemize{
  \item \code{gd} calculates the Generational Distance (GD).
  \item \code{igd} calculates the Inverse Generational Distance (IGD).
  \item \code{gdp/igdp} compute the (I)GD+ indicator [3].
  \item \code{ahd} calculates the Average Hausdorff Distance.
}
}
\details{
The Generational Distance (GD) measures the distance of a point set
\eqn{X = \{r_1, \ldots, r_{|X|}\}}, e.g., a Pareto-front approximation, to
a reference set \eqn{R = \{r_1, \ldots, r_{|R|}\}}. Then GD is defined as
\deqn{
 GD_p(A, R) = \frac{1}{|X|} \left(\sum_{i=1}^{|X|} d_i^p\right)^{1/p}
}
where \eqn{d_i} is the Euclidean distance of point \eqn{x_i \in X} to
its nearest neigbor point in \eqn{R}. The Inverted Generational Distance works
the other way around, i.e.,
\deqn{
 IGD_p(A, R) = \frac{1}{|R|} \left(\sum_{i=1}^{|R|} \hat{d}_i^p\right)^{1/p}
}
where \eqn{\hat{d}_i} is the respective nearest neighbor distance of \eqn{r_i}
to any point in \eqn{X}. Put differently, \eqn{IGD_p(A, R) = GD_p(R, A)}.
Functions \code{gd} and \code{igd} calcute these versions.

Schütze et al. [2] proposed a slight modification:
\deqn{
 GD_p(A, R) = \left(\frac{1}{|X|} \sum_{i=1}^{|X|} d_i^p\right)^{1/p}
}
where the average is taken before the power operation. \eqn{IGD_p} is apdated
analogeously. This versions are calclated by \code{gd} and \code{igd} if the argument
\code{modified} is set to \code{TRUE}.

Ishibushi et al. [3] proposed another modification which works on the
formulation by Schütze et al. (see above). They modified the distance
calculation:
\deqn{
 GD_p^{+}(A, R) = \left(\frac{1}{|X|} \sum_{i=1}^{|X|} d^{+^p}_i\right)^{1/p}
}
where \eqn{d_i^{+} = \max\{x_i, z_i\}}. This version can be calculated
with the function \code{gdp} (the trailing \emph{p} stands for \dQuote{plus}).

Eventuelly, the function \code{ahd} calculates the Average Hausdorff Distance [2]
which combines GD and IGD and is defined as
\deqn{
 \Delta_p(A, R) = \max\{GD_p(A, R), IGD_p(A, R)\}.
}
By default, \code{ahd} uses the modified versions of \eqn{GD} and \eqn{IGD}
respectively (see argument \code{modified}).

IGDX [4] is a meaasure for decision space diversity. This is simply IGD;
however, the input consists of the non-dominated solutions in decision space
rather in objective space. Naturally, all implemented functions can be used as
an \dQuote{*X} version.
}
\references{
[1] David A. Van Veldhuizen and David A. Van Veldhuizen. Multiobjective
evolutionary algorithms: classifications, analyses, and new innovations.
Technical Report, Evolutionary Computation, 1999.

[2] Schütze, O., Esquivel, X.,Lara,A. ,Coello, C.A.C.: Using the averaged
Hausdorff distance as a performance measure in evolutionary multiobjective
optimization. IEEE Transactions on Evolutionary Computation 16, 504–522 (2012).

[3] Hisao Ishibuchi, Hiroyuki Masuda, Yuki Tanigaki, and Yusuke Nojima.
Modified distance calculation in generational distance and inverted generational
distance. In António Gaspar-Cunha, Carlos Henggeler Antunes, and Carlos
Coello Coello, editors, Evolutionary Multi-Criterion Optimization, 110–125.
Cham, 2015. Springer International Publishing.

[4] O. Schütze, M. Vasile, and C. A. C. Coello, Computing the Set of
Epsilon-Efficient Solutions in Multiobjective Space Mission Design,
}
\seealso{
Other multi-objective performance indicators: 
\code{\link{cov}()},
\code{\link{df_get_indicators}()},
\code{\link{eps}()},
\code{\link{hv}()},
\code{\link{os}()},
\code{\link{r1}()},
\code{\link{rse}()}
}
\concept{multi-objective performance indicators}
\keyword{optimize}
