% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_hv.R
\name{hv}
\alias{hv}
\alias{hv_contr}
\title{Dominated hypervolume (contribution).}
\usage{
hv(x, ref.point = NULL, offset = 1, ...)

hv_contr(x, ref.point = NULL, offset = 1)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Matrix of points (column-wise).}

\item{ref.point}{[\code{numeric} | \code{NULL}]\cr
Reference point.
Set to the maximum in each dimension by default if not provided.}

\item{offset}{[\code{numeric(1)}]\cr
Offset to be added to each component of the reference point only in the case
where no reference is provided and one is calculated automatically.
Default is 1.}

\item{...}{[any]\cr
Not used at the moment.}
}
\value{
[\code{numeric(1)}] Dominated hypervolume in the case of
 \code{hv} and the dominated hypervolume contributions
 for each point in the case of \code{hv_contr}.
}
\description{
The function \code{hv} computes the dominated hypervolume of a set of points
given a reference set whereby \code{hv_contr} computes the hypervolume contribution
of each point.

If no reference point is given the nadir point of the set \code{x} is
determined and a positive offset with default 1 is added. This is to ensure
that the reference point is dominated by all of the points in the reference set.
}
\note{
Keep in mind that this function assumes all objectives to be minimized.
In case at least one objective is to be maximized the matrix \code{x} needs
to be transformed accordingly in advance.
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other multi-objective performance indicators: 
\code{\link{cov}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{os}()}
}
\concept{mootools}
\concept{multi-objective performance indicators}
