% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_os.R
\name{os}
\alias{os}
\title{Overall (Pareto) spread}
\usage{
os(x, np = NULL, ip = NULL)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Point set in column major format.}

\item{np}{[\code{numeric}]\cr
Nadir point or approximtion of it.}

\item{ip}{[\code{numeric}]\cr
Ideal point or approximtion of it.}
}
\value{
[\code{numeric(1)}] Scalar indicator value.
}
\description{
A simple multi-objective performance indicator proposed in [1].
Given a point set\eqn{X = \{x_1, \ldots, x_{n}\}} of \eqn{m} objectives,
the ideal point \eqn{I} (or an approximation) and the nadir point \eqn{N}
(or a approximation) it is defined as
\deqn{
  OS(X) = \prod_{i=1}^{m} \frac{\max_{x \in X} x_i - \min_{x \in X} x_i}{N_i - I_i}.
}
This indicator only captures the extend of the Pareto-front approximation.
}
\references{
[1] J. Wu, S. Azarm, Metrics for quality assessment of a multiobjective design
optimization solution set, Journal of Mechanical Design 123 (1) (2001) 18–25.
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other multi-objective performance indicators: 
\code{\link{cov}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{hv}()}
}
\concept{mootools}
\concept{multi-objective performance indicators}
\keyword{optimize}
