% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_eaf.R
\name{plot_eaf}
\alias{plot_eaf}
\title{Empirical attainment function (EAF)}
\usage{
plot_eaf(df, obj.cols, percentiles = c(0, 50, 75, 100))
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{prob} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective functions.
Default is \code{c("f1", "f2")}.}

\item{percentiles}{[\code{numeric}]\cr
Percentiles of the EAF that will be plotted as attainment surfaces.}
}
\value{
[\code{\link[ggplot2]{ggplot}}] ggplot object.
}
\description{
...
}
\references{
[1] V. Grunert da Fonseca and C. M. Fonseca, "The attainment-function approach
to stochastic multiobjective optimizer assessment and comparison," in
Experimental Methods for the Analysis of Optimization Algorithms
(T. Bartz-Beielstein, M. Chiarandini, L. Paquete, and M. Preuss, eds.),
ch. 5, pp. 103-130, Springer Berlin Heidelberg, 2010.

[2] Manuel López-Ibáñez, Luís Paquete, and Thomas Stützle. Exploratory Analysis
of Stochastic Local Search Algorithms in Biobjective Optimization. In T.
Bartz-Beielstein, M. Chiarandini, L. Paquete, and M. Preuss, editors,
Experimental Methods for the Analysis of Optimization Algorithms,
pages 209–222. Springer, Berlin, Germany, 2010. doi: 10.1007/978-3-642-02538-9_9
}
\seealso{
Other mootools: 
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other visualization: 
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()}
}
\concept{mootools}
\concept{visualization}
