% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap}
\usage{
plot_heatmap(df, obj.cols = c("y1", "y2"))
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{problem} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective function values.
Default is \code{c("y1", "y2")}.}
}
\value{
[\code{\link[ggplot2]{ggplot}}] ggplot object.
}
\description{
Visualize a Pareto-front approximation set by means of a heatmap.
Here, on the \eqn{x}-axis the objectives are given in the order provided by
\code{obj.cols}. The \eqn{y}-axis shows the different solutions. The color of the
cells indicates the objective value of the respective (objective, solution)-pair.
Works with any kind of objectives.
}
\references{
[1] T. Tušar and B. Filipič, Visualization of Pareto Front Approximations in
Evolutionary Multiobjective Optimization: A Critical Review and the Prosection
Method, in IEEE Transactions on Evolutionary Computation, vol. 19, no. 2,
pp. 225-245, April 2015, doi: 10.1109/TEVC.2014.2313407.
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{df_get_indicators}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{r1}()},
\code{\link{rse}()},
\code{\link{set_dominates}()},
\code{\link{solow_polasky}()},
\code{\link{which_dominated}()}

Other visualization: 
\code{\link{plot_eaf}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()}
}
\concept{mootools}
\concept{visualization}
