% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_pcp.R
\name{plot_pcp}
\alias{plot_pcp}
\title{Parallel coordinate plot (PCP)}
\usage{
plot_pcp(df, obj.cols = c("y1", "y2"))
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{prob} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective functions.
Default is \code{c("f1", "f2")}.}
}
\value{
[\code{\link[ggplot2]{ggplot}}] ggplot object.
}
\description{
Visualizes a Pareto-front approximation set with parallel coordinate
plots. The \eqn{x}-axis shows the objectives as an ordered factor where the order
is determined by the \code{obj.cols} argument. On the \eqn{y}-axis objective values
are shown colored by solution as lines. Suitable for any number of objectives.
}
\references{
[1] T. Tušar and B. Filipič, Visualization of Pareto Front Approximations in
Evolutionary Multiobjective Optimization: A Critical Review and the Prosection
Method, in IEEE Transactions on Evolutionary Computation, vol. 19, no. 2,
pp. 225-245, April 2015, doi: 10.1109/TEVC.2014.2313407.
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other visualization: 
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()}
}
\concept{mootools}
\concept{visualization}
