% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Radar plot}
\usage{
plot_radar(df, obj.cols = c("y1", "y2", "y3"))
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data.frame with columns at least \code{obj.cols}, \dQuote{problem} and \dQuote{algorithm}.}

\item{obj.cols}{[\code{character(>= 2)}]\cr
Column names of the objective function values.
Default is \code{c("y1", "y2")}.}
}
\value{
[\code{\link[ggplot2]{ggplot}}] ggplot object.
}
\description{
A radar plot or spider plot is an elegant method to visualize
solution quality of solutions with more than two objectives. Essentially it
shows multiple numeric values starting from the same origin in a circular
two-dimensional representation

Users who are into gaming might know the kind of visualization from games like
Pro Evolution Soccer, NHL or most likely any kind of sport games. Here, the
individual strength and weaknesses of players (speed, tequniqual skills,
fitness etc.) are often visualized by means of radar plots.

The drawback of this method that it is difficult to compare many solutions
or - even worse - many solutions from different algorithms to get the big picture.
}
\examples{
data(mtcars)

# scale to [0, 1]
tbl = as.data.frame(scale(mtcars[1:6, ], center = FALSE, scale = TRUE))
tbl$car = rownames(tbl)

\dontrun{
# 6 cars in one plot
g = plot_radar(tbl, c("mpg", "cyl", "hp"))
print(g)

# split by car (-> 6 plots)
print(g + facet_wrap(car ~ .))
}
}
\references{
[1] T. Tušar and B. Filipič, Visualization of Pareto Front Approximations in
Evolutionary Multiobjective Optimization: A Critical Review and the Prosection
Method, in IEEE Transactions on Evolutionary Computation, vol. 19, no. 2,
pp. 225-245, April 2015, doi: 10.1109/TEVC.2014.2313407.
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{df_get_indicators}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{r1}()},
\code{\link{rse}()},
\code{\link{set_dominates}()},
\code{\link{solow_polasky}()},
\code{\link{which_dominated}()}

Other visualization: 
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()}
}
\concept{mootools}
\concept{visualization}
