% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_reference_points.R
\name{get_nadir}
\alias{get_nadir}
\alias{get_ideal}
\title{Reference point approximations.}
\usage{
get_nadir(..., sets = NULL)

get_ideal(..., sets = NULL)
}
\arguments{
\item{...}{[\code{matrix}]\cr
Arbirary number of matrizes.}

\item{sets}{[\code{list}]\cr
List of matrizes. This is an alternative way of passing the sets. Can be used
exclusively or combined with \code{...}.}
}
\value{
[\code{numeric}] Reference point.
}
\description{
Helper functions to compute nadir or ideal point from sets of
points, e.g., multiple approximation sets.
}
\seealso{
Other mootools: 
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}
}
\concept{mootools}
