% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_rse.R
\name{rse}
\alias{rse}
\title{Riesz \eqn{s}-energy}
\usage{
rse(x, s)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Point set in column major format.}

\item{s}{[\code{numeric}]\cr
Parameter \eqn{s \geq 0} (see description).}
}
\value{
[\code{numeric(1)}] Scalar indicator value.
}
\description{
The Riesz \eqn{s}-energy is designed as a measure for the evenness of a set
of points \eqn{X = \{x_1, \ldots, x_{|X|}\}}. It is formally defined as
\deqn{
  R_s(X) = \sum_{x \in X} \sum_{y \in X, y \neq x} k_s(x, y)
}
where function
\deqn{
  k_s(x, y) =
  \begin{cases}
    d(x, y)^{-s}, s > 0 \\
    -\log(d(x, y)), s = 0
  \end{cases}
}
is the so-called Riesz \eqn{s}-kernel and \eqn{d(x,y)} is the Euclidean
distance between \eqn{x} and \eqn{y}. The parameter \eqn{s \geq 0} steers
the desired degree of uniformity of the distribution with increasing
emphasis on uniformity for \eqn{s \to \infty}. See [1] for an application
of the Riesz \eqn{s}-energy in multi-objective evolutionary optimization
and [2, 3] for a mathematically rigorous introduction into the general idea.
}
\references{
[1] J. G. Falcón-Cardona, H. Ishibuchi and C. A. C. Coello, Riesz s-energy-based
Reference Sets for Multi-Objective optimization," 2020 IEEE Congress on
Evolutionary Computation (CEC), 2020, pp. 1-8, doi: 10.1109/CEC48606.2020.9185833.

[2] D.P.HardinandE.B.Saff, MinimalRieszenergypointconfigurations for rectifiable
d-dimensional manifolds, Advances in Mathematics, vol. 193, no. 1, pp. 174–204, 2005.

[3] D. P. Hardin and E. B. Saff, Discretizing Manifolds via Minimum Energy Points,
Notices of the AMS, vol. 51, no. 10, pp. 1186–1194, 2004.
}
\seealso{
Other mootools: 
\code{\link{cov}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{r1}()},
\code{\link{set_dominates}()},
\code{\link{which_dominated}()}

Other multi-objective performance indicators: 
\code{\link{cov}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{hv}()},
\code{\link{os}()},
\code{\link{r1}()}
}
\concept{mootools}
\concept{multi-objective performance indicators}
\keyword{optimize}
