% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moo_domination.R
\name{set_dominates}
\alias{set_dominates}
\title{Check if one set is better than another.}
\usage{
set_dominates(x, y)
}
\arguments{
\item{x}{[\code{matrix}]\cr
First set of points.}

\item{y}{[\code{matrix}]\cr
Second set of points.}
}
\value{
[\code{logical(1)}]
}
\description{
The function checks, whether every point from the second set of points
is dominated by at least one point from the first set.
}
\examples{
x = matrix(c(1, 1, 2, 2, 1, 3), byrow = FALSE, ncol = 3L)
y = matrix(c(3, 4, 2, 2, 5, 6, 7, 7), byrow = FALSE, ncol = 4L)
set_dominates(x, y)
set_dominates(y, x)
}
\seealso{
Other pareto-dominance: 
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{which_dominated}()}

Other mootools: 
\code{\link{cov}()},
\code{\link{df_get_indicators}()},
\code{\link{dominated}()},
\code{\link{dominates}()},
\code{\link{eps}()},
\code{\link{gd}()},
\code{\link{get_nadir}()},
\code{\link{hv}()},
\code{\link{nds}()},
\code{\link{os}()},
\code{\link{plot_eaf}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_pcp}()},
\code{\link{plot_radar}()},
\code{\link{plot_scatter2d}()},
\code{\link{plot_scatter3d}()},
\code{\link{r1}()},
\code{\link{rse}()},
\code{\link{solow_polasky}()},
\code{\link{which_dominated}()}
}
\concept{mootools}
\concept{pareto-dominance}
